fix: move `embed_text_values` to use timeseriesflattener without text dependencies installed
